Remove travis.yml altogether
This project requires Microsoft.VSSDK tools, which are not compatible with
OSX or Linux environments, which are the only currently viable options to
run the Travis pipeline on. Windows does not support csharp projects yet.

In the future, I may work on setting up a custom environment on travis, or
use an Azure pipeline instead.